Level DB: use cache to avoid unecessary S3 requests
The use of db for cache might reduce the number of requests necessary to
upload a site to S3. This shall speed up the deployment and maybe reduce
a few cents in the bill of huge sites owners. For small sites the number
of requests does not matter since it shall fall into Amazon free tier.

Signed-off-by: Josenivaldo Benito Jr <jrbenito@benito.qsl.br>