Switch s3-sync to s3-sync-aws
s3-sync is not mantained and it uses knox to S3 acess. However knox does
not work for new S3 regions like Frankfurt. The s3-sync-aws is a fork
from original s3-sync and switches from knox to AWS SDK, a better
approach for compatability.

Before fork s3-sync to s3-sync-aws, author offered a pull request to
former, the PR is still active and no forecast to be merged. So this
proposes a switching from s3-sync to s3-sync-aws to address new S3
regions issues and stays aligned with AWS new features.

TODO: check if cloudfront dependency could be dropped in favor of AWS
SDK to manage cloudfront invalidations

Signed-off-by: Josenivaldo Benito Jr <jrbenito@benito.qsl.br>