Pass headers: make possible to pass custom headers ahead to s3-sync
S3-sync has the ability to set custom headers to be passed to objects
and saved as metadata. S3-sync itself use this ability to save a md5sum
hash of each object in a custom meta header.

This patch includes ability to hexo-deployer-s3-cloudfront to pass
custom headers (for all objects at once) to s3-sync save them as headers
and metadata. Main inspiration is to be able to pass CacheControl and
set Cache-Control header to objects since this avoids "Leverage Browse
Caching" warnings in most page speed tests. However, any metadata can be
passed to objects as examples provided shows.

Signed-off-by: Josenivaldo Benito Jr <jrbenito@benito.qsl.br>